Create files util in order to re-use getting files and offset index code in wheel component and profile action
