+++ src/actions/profile.js
@@ -4,20 +4,20 @@ import {
   UPDATE_PROFILE_ASSET_COLOR,
   UPDATE_PROFILE_ASSET_VISIBILITY
 } from '../constants/profile';
+import { getIndexByOffset, getFiles } from '../utils/files';
 
 export const updateProfileAssetFileIndex = offset => (dispatch, getState) => {
   const { assets, profile } = getState();
-  const currentAsset = assets.items[assets.current.asset];
-  const { asset, type, color } = assets.current;
-  const files = currentAsset.types[type].colors[color].files;
-  const length = files.length;
+  const files = getFiles(assets);
+  const fileIndex = profile[assets.current.asset].fileIndex;
+  const newIndex = getIndexByOffset(files.length, fileIndex, offset);
   dispatch({
     type: UPDATE_PROFILE_ASSET,
     asset: assets.current.asset,
     payload: {
-      type,
-      color,
-      fileIndex: (length + profile[assets.current.asset].fileIndex + offset) % length,
+      type: files[newIndex].type,
+      color: files[newIndex].color,
+      fileIndex: newIndex,
       visible: true
     }
   });
